pyproject.toml
[project]
name = "neural-cssr"
version = "0.1.0"
description = "Neural CSSR implementation"
requires-python = ">=3.8"
dependencies = [
    "torch",
    "numpy", 
    "pyyaml",
    "scipy"
]